Factor out OS-specific function from startup.go.
