+++ pkg/startup/startup.go
@@ -55,14 +55,6 @@ const (
 	AUTODETECTION_METHOD_SKIP_INTERFACE = "skip-interface="
 )
 
-// Default interfaces to exclude for any logic following the first-found
-// auto detect IP method
-var DEFAULT_INTERFACES_TO_EXCLUDE []string = []string{
-	"docker.*", "cbr.*", "dummy.*",
-	"virbr.*", "lxcbr.*", "veth.*", "lo",
-	"cali.*", "tunl.*", "flannel.*", "kube-ipvs.*",
-}
-
 // Version string, set during build.
 var VERSION string
 
@@ -260,14 +252,14 @@ func determineNodeName() string {
 
 	// Determine the name of this node.  Precedence is:
 	// -  NODENAME
-	// -  Value stored in /var/lib/calico/nodename
+	// -  Value stored in our nodename file.
 	// -  HOSTNAME (lowercase)
 	// -  os.Hostname (lowercase).
 	// We use the names.Hostname which lowercases and trims the name.
 	if nodeName = strings.TrimSpace(os.Getenv("NODENAME")); nodeName != "" {
 		log.Infof("Using NODENAME environment for node name")
 	} else if nodeName = nodenameFromFile(); nodeName != "" {
-		log.Info("Using stored node name from /var/lib/calico/nodename")
+		log.Info("Using stored node name from " + nodenameFileName())
 	} else if nodeName = strings.ToLower(strings.TrimSpace(os.Getenv("HOSTNAME"))); nodeName != "" {
 		log.Infof("Using HOSTNAME environment (lowercase) for node name")
 	} else if nodeName, err = names.Hostname(); err != nil {
@@ -282,17 +274,26 @@ func determineNodeName() string {
 	return nodeName
 }
 
-// nodenameFromFile reads the /var/lib/calico/nodename file if it exists and
+func nodenameFileName() string {
+	fn := os.Getenv("CALICO_NODENAME_FILE")
+	if fn == "" {
+		return defaultNodenameFile
+	}
+	return fn
+}
+
+// nodenameFromFile reads the nodename file if it exists and
 // returns the nodename within.
 func nodenameFromFile() string {
-	data, err := ioutil.ReadFile("/var/lib/calico/nodename")
+	filename := nodenameFileName()
+	data, err := ioutil.ReadFile(filename)
 	if err != nil {
 		if os.IsNotExist(err) {
 			// File doesn't exist, return empty string.
-			log.Debug("File does not exist: /var/lib/calico/nodename")
+			log.Debug("File does not exist: " + filename)
 			return ""
 		}
-		log.WithError(err).Error("Failed to read /var/lib/calico/nodename")
+		log.WithError(err).Error("Failed to read " + filename)
 		terminate()
 	}
 	return string(data)
@@ -329,11 +330,12 @@ func waitForConnection(ctx context.Context, c client.Interface) {
 
 // writeNodeConfig writes out the this node's configuration to disk for use by other components.
 // Specifically, it creates:
-// - /var/lib/calico/nodename - used to persist the determined node name to disk for future use.
+// - nodenameFileName() - used to persist the determined node name to disk for future use.
 func writeNodeConfig(nodeName string) {
-	log.Debugf("Writing %s to /var/lib/calico/nodename", nodeName)
-	if err := ioutil.WriteFile("/var/lib/calico/nodename", []byte(nodeName), 0644); err != nil {
-		log.WithError(err).Error("Unable to write to /var/lib/calico/nodename")
+	filename := nodenameFileName()
+	log.Debugf("Writing %s to "+filename, nodeName)
+	if err := ioutil.WriteFile(filename, []byte(nodeName), 0644); err != nil {
+		log.WithError(err).Error("Unable to write to " + filename)
 		terminate()
 	}
 }
@@ -735,24 +737,6 @@ func configureIPPools(ctx context.Context, client client.Interface) {
 	}
 }
 
-// ipv6Supported returns true if IPv6 is supported on this platform.  This performs
-// a check on the appropriate Felix parameter and if supported also performs a
-// simplistic check of /proc/sys/net/ipv6 (since platforms that do not have IPv6
-// compiled in will not have this entry).
-func ipv6Supported() bool {
-	// First check if Felix param is false
-	IPv6isSupported := evaluateENVBool("FELIX_IPV6SUPPORT", true)
-	if !IPv6isSupported {
-		return false
-	}
-
-	// If supported, then also check /proc/sys/net/ipv6.
-	_, err := os.Stat("/proc/sys/net/ipv6")
-	supported := (err == nil)
-	log.Infof("IPv6 supported on this platform: %v", supported)
-	return supported
-}
-
 // createIPPool creates an IP pool using the specified CIDR.  This
 // method is a no-op if the pool already exists.
 func createIPPool(ctx context.Context, client client.Interface, cidr *cnet.IPNet, poolName, ipipModeName string, isNATOutgoingEnabled bool) {
@@ -876,52 +860,6 @@ func checkConflictingNodes(ctx context.Context, client client.Interface, node *a
 		}
 	}
 	return
-}
-
-// Checks that the filesystem is as expected and fix it if possible
-func ensureFilesystemAsExpected() {
-	// BIRD requires the /var/run/calico directory in order to provide status
-	// information over the control socket, but other backends do not
-	// need this check.
-	if strings.ToLower(os.Getenv("CALICO_NETWORKING_BACKEND")) == "bird" {
-		runDir := "/var/run/calico"
-		// Check if directory already exists
-		if _, err := os.Stat(runDir); err != nil {
-			// Create the runDir
-			if err = os.MkdirAll(runDir, os.ModeDir); err != nil {
-				log.Errorf("Unable to create '%s'", runDir)
-				terminate()
-			}
-			log.Warnf("Expected %s to be mounted into the container but it wasn't present. 'calicoctl node status' may provide incomplete status information", runDir)
-		}
-	}
-
-	// Make sure the /var/lib/calico directory exists.
-	libDir := "/var/lib/calico"
-	// Check if directory already exists
-	if _, err := os.Stat(libDir); err != nil {
-		// Create the libDir
-		if err = os.MkdirAll(libDir, os.ModeDir); err != nil {
-			log.Errorf("Unable to create '%s'", libDir)
-			terminate()
-		}
-		log.Warnf("Expected %s to be mounted into the container but it wasn't present. Node name may not be detected properly", libDir)
-	}
-
-	// Ensure the log directory exists but only if logging to file is enabled.
-	if strings.ToLower(os.Getenv("CALICO_DISABLE_FILE_LOGGING")) != "true" {
-		logDir := "/var/log/calico"
-		// Check if directory already exists
-		if _, err := os.Stat(logDir); err != nil {
-			// Create the logDir
-			if err = os.MkdirAll(logDir, os.ModeDir); err != nil {
-				log.Errorf("Unable to create '%s'", logDir)
-				terminate()
-			}
-			log.Warnf("Expected %s to be mounted into the container but it wasn't present. 'calicoctl node diags' will not be able to collect calico/node logs", logDir)
-		}
-	}
-
 }
 
 // ensureDefaultConfig ensures all of the required default settings are
